+++ c_src/Astar.h
@@ -31,7 +31,7 @@
 
 typedef enum {UNKNOWN, OPEN, CLOSED} astar_status;
 
-int16_t astar(int *map, int16_t width, int16_t height, int16_t *route, int16_t start, int16_t destination);
+int16_t astar(int16_t *map, int16_t width, int16_t height, int16_t *route, int16_t start, int16_t destination);
 
 static void init_astar(int16_t *open, int16_t *closed, int16_t *parent, int16_t *g_cost, int16_t size);
 
@@ -41,7 +41,7 @@ static int16_t next_current(int16_t *open, int16_t *g_cost, int16_t destination,
 
 static void add_closed(int16_t *closed, int16_t location);
 
-static uint8_t add_open(int *map, int16_t *open, int16_t *closed, int16_t *g_cost, int16_t *parent, int16_t current);
+static uint8_t add_open(int16_t *map, int16_t *open, int16_t *closed, int16_t *g_cost, int16_t *parent, int16_t current);
 
 static uint8_t traceback(int16_t *parent, int16_t *traceback, int16_t destination, int16_t start);
